Rename files
Rename files based on suggestions from @markbao
