+++ src/index.tsx
@@ -1,4 +1,4 @@
 import { h, render } from "preact";
-import { App } from "./app";
+import { App } from "./App";
 
 render(<App />, document.body);
