+++ src/pages/EncounterBuilder.tsx
