+++ src/pages/EnvironmentDetail.tsx
