+++ src/pages/MonsterDetail.tsx
