+++ src/pages/SourceDetail.tsx
